ec2-zimage-turbo/requirements.txt
torch>=2.0.0
diffusers>=0.32.0
transformers>=4.30.0
accelerate>=0.20.0
flask>=2.3.0
pillow>=10.0.0
safetensors>=0.4.0
sentencepiece>=0.1.99
